pres

+++ back/logic/gameManagers/OneVOneOnlineGameManager.js
@@ -3,7 +3,12 @@ const { Game } = require("../../../front/utils/game.js");
 const { SocketMapper } = require("../../socket/socketMapper.js");
 const { SocketSender } = require("../../socket/socketSender.js");
 
-const { saveElo,winAGameAchievement,loseAGameAchievement,placeAWallAchievement } = require("../../mongoDB/mongoManager");
+const {
+  saveElo,
+  winAGameAchievement,
+  loseAGameAchievement,
+  placeAWallAchievement,
+} = require("../../mongoDB/mongoManager");
 const { GameManagerMapper } = require("./gameManagerMapper");
 const { AchievementsManager } = require("../../social/achievements");
 
@@ -47,6 +52,7 @@ class OneVOneOnlineGameManager {
   }
 
   playerWon(playerNumber) {
+    if (this.game === null) return;
     GameManagerMapper.removeOnlineGameManagerByUserId(this.idClient1);
     GameManagerMapper.removeOnlineGameManagerByUserId(this.idClient2);
 
@@ -68,14 +74,12 @@ class OneVOneOnlineGameManager {
     saveElo(this.idClient2, this.eloClient2 + deltaClient2);
 
     const wallsClient1 = 10 - this.game.players[0].nbWalls;
-    placeAWallAchievement(this.idClient1,wallsClient1);
-
-    
+    placeAWallAchievement(this.idClient1, wallsClient1);
 
     const wallsClient2 = 10 - this.game.players[1].nbWalls;
-    placeAWallAchievement(this.idClient2,wallsClient2);
+    placeAWallAchievement(this.idClient2, wallsClient2);
 
-    switch (playerNumber){
+    switch (playerNumber) {
       case 1:
         winAGameAchievement(this.idClient1);
         loseAGameAchievement(this.idClient2);
